+++ packages/mongo/collection.js
@@ -55,20 +55,20 @@ Mongo.Collection = function Collection(name, options) {
     transform: null,
     _driver: undefined,
     _preventAutopublish: false,
-      ...options,
+    ...options,
   };
 
   switch (options.idGeneration) {
   case 'MONGO':
     this._makeNewID = function () {
-      var src = name ? DDP.randomStream('/collection/' + name) : Random.insecure;
+      const src = name ? DDP.randomStream('/collection/' + name) : Random.insecure;
       return new Mongo.ObjectID(src.hexString(24));
     };
     break;
   case 'STRING':
   default:
     this._makeNewID = function () {
-      var src = name ? DDP.randomStream('/collection/' + name) : Random.insecure;
+      const src = name ? DDP.randomStream('/collection/' + name) : Random.insecure;
       return src.id();
     };
     break;
@@ -175,8 +175,8 @@ Object.assign(Mongo.Collection.prototype, {
       // Apply an update.
       // XXX better specify this interface (not in terms of a wire message)?
       update(msg) {
-        var mongoId = MongoID.idParse(msg.id);
-        var doc = self._collection._docs.get(mongoId);
+        const mongoId = MongoID.idParse(msg.id);
+        const doc = self._collection._docs.get(mongoId);
 
         // Is this a "replace the whole doc" message coming from the quiescence
         // of method writes to an object? (Note that 'undefined' is a valid
@@ -207,7 +207,7 @@ Object.assign(Mongo.Collection.prototype, {
             throw new Error("Expected to find a document to change");
           const keys = Object.keys(msg.fields);
           if (keys.length > 0) {
-            var modifier = {};
+            const modifier = {};
             keys.forEach(key => {
               const value = msg.fields[key];
               if (EJSON.equals(doc[key], value)) {
@@ -288,7 +288,7 @@ Object.assign(Mongo.Collection.prototype, {
   },
 
   _getFindOptions(args) {
-    var self = this;
+    const self = this;
     if (args.length < 2) {
       return { transform: self._transform };
     } else {
@@ -362,7 +362,7 @@ Object.assign(Mongo.Collection.prototype, {
 
 Object.assign(Mongo.Collection, {
   _publishCursor(cursor, sub, collection) {
-    var observeHandle = cursor.observeChanges({
+    const observeHandle = cursor.observeChanges({
       added: function (id, fields) {
         sub.added(collection, id, fields);
       },
@@ -446,7 +446,7 @@ Object.assign(Mongo.Collection.prototype, {
   // off.
 
   /**
-   * @summary Insert a document in the collection.  Returns its unique _id.
+   * @summary Insert a document into the collection.  Returns its unique _id.
    * @locus Anywhere
    * @method  insert
    * @memberof Mongo.Collection
@@ -455,51 +455,76 @@ Object.assign(Mongo.Collection.prototype, {
    * @param {Function} [callback] Optional.  If present, called with an error object as the first argument and, if no error, the _id as the second.
    */
   insert(doc, callback) {
+    const isBulkInsert = Array.isArray(doc);
+
     // Make sure we were passed a document to insert
     if (!doc) {
       throw new Error("insert requires an argument");
     }
 
-    // Make a shallow clone of the document, preserving its prototype.
-    doc = Object.create(
-      Object.getPrototypeOf(doc),
-      Object.getOwnPropertyDescriptors(doc)
-    );
+    const checkId = document => {
+      // Make a shallow clone of the document, preserving its prototype.
+      const dct = Object.create(
+        Object.getPrototypeOf(document),
+        Object.getOwnPropertyDescriptors(document)
+      );
+      if (dct._id || typeof dct._id === 'string') {
+        // If _id is present in the doc check that it is in the correct format.
+        if (! dct._id ||
+          ! ((typeof dct._id === 'string' && !!dct._id.trim()) ||
+            doc._id instanceof Mongo.ObjectID)) {
+          throw new Error(
+            "Meteor requires document _id fields to be non-empty strings or ObjectIDs");
+        }
+      } else {
+        // We need to generate the _id for the document
+        let generateId = true;
 
-    if ('_id' in doc) {
-      if (! doc._id ||
-          ! (typeof doc._id === 'string' ||
-             doc._id instanceof Mongo.ObjectID)) {
-        throw new Error(
-          "Meteor requires document _id fields to be non-empty strings or ObjectIDs");
-      }
-    } else {
-      let generateId = true;
+        // Don't generate the id if we're the client and the 'outermost' call
+        // This optimization saves us passing both the randomSeed and the id
+        // Passing both is redundant.
+        if (this._isRemoteCollection()) {
+          const enclosing = DDP._CurrentMethodInvocation.get();
+          if (!enclosing) {
+            generateId = false;
+          }
+        }
 
-      // Don't generate the id if we're the client and the 'outermost' call
-      // This optimization saves us passing both the randomSeed and the id
-      // Passing both is redundant.
-      if (this._isRemoteCollection()) {
-        const enclosing = DDP._CurrentMethodInvocation.get();
-        if (!enclosing) {
-          generateId = false;
+        if (generateId) {
+          dct._id = this._makeNewID();
         }
       }
+      return dct;
+    }
 
-      if (generateId) {
-        doc._id = this._makeNewID();
-      }
+    if (isBulkInsert) {
+      // Clone the array
+      const insertDocs = doc.map(d => {
+        if (!d) return null;
+        return checkId(d);
+      }).filter(Boolean);
+      doc = insertDocs;
+    } else {
+      doc = checkId(doc);
     }
 
     // On inserts, always return the id that we generated; on all other
     // operations, just return the result from the collection.
-    var chooseReturnValueFromCollectionResult = function (result) {
+    const chooseReturnValueFromCollectionResult = function (result, minimongo = false) {
+      if (minimongo) return result;
+      if (isBulkInsert) {
+        if (result.insertedIds) {
+          return Object.values(result.insertedIds);
+        }
+        return [];
+      }
+
       if (doc._id) {
         return doc._id;
       }
 
       // XXX what is this for??
-      // It's some iteraction between the callback to _callMutatorMethod and
+      // It's some interaction between the callback to _callMutatorMethod and
       // the return value conversion
       doc._id = result;
 
@@ -510,8 +535,8 @@ Object.assign(Mongo.Collection.prototype, {
       callback, chooseReturnValueFromCollectionResult);
 
     if (this._isRemoteCollection()) {
-      const result = this._callMutatorMethod("insert", [doc], wrappedCallback);
-      return chooseReturnValueFromCollectionResult(result);
+      const result = this._callMutatorMethod("insert", isBulkInsert ? doc : [doc], wrappedCallback);
+      return chooseReturnValueFromCollectionResult(result, true);
     }
 
     // it's my collection.  descend into the collection object
@@ -665,41 +690,41 @@ Object.assign(Mongo.Collection.prototype, {
   // We'll actually design an index API later. For now, we just pass through to
   // Mongo's, but make it synchronous.
   _ensureIndex(index, options) {
-    var self = this;
+    const self = this;
     if (!self._collection._ensureIndex)
       throw new Error("Can only call _ensureIndex on server collections");
     self._collection._ensureIndex(index, options);
   },
 
   _dropIndex(index) {
-    var self = this;
+    const self = this;
     if (!self._collection._dropIndex)
       throw new Error("Can only call _dropIndex on server collections");
     self._collection._dropIndex(index);
   },
 
   _dropCollection() {
-    var self = this;
+    const self = this;
     if (!self._collection.dropCollection)
       throw new Error("Can only call _dropCollection on server collections");
     self._collection.dropCollection();
   },
 
   _createCappedCollection(byteSize, maxDocuments) {
-    var self = this;
+    const self = this;
     if (!self._collection._createCappedCollection)
       throw new Error("Can only call _createCappedCollection on server collections");
     self._collection._createCappedCollection(byteSize, maxDocuments);
   },
 
   /**
-   * @summary Returns the [`Collection`](http://mongodb.github.io/node-mongodb-native/3.0/api/Collection.html) object corresponding to this collection from the [npm `mongodb` driver module](https://www.npmjs.com/package/mongodb) which is wrapped by `Mongo.Collection`.
+   * @summary Returns the [`Collection`](http://mongodb.github.io/node-mongodb-native/3.6/api/Collection.html) object corresponding to this collection from the [npm `mongodb` driver module](https://www.npmjs.com/package/mongodb) which is wrapped by `Mongo.Collection`.
    * @locus Server
    * @memberof Mongo.Collection
    * @instance
    */
   rawCollection() {
-    var self = this;
+    const self = this;
     if (! self._collection.rawCollection) {
       throw new Error("Can only call rawCollection on server collections");
     }
@@ -707,7 +732,7 @@ Object.assign(Mongo.Collection.prototype, {
   },
 
   /**
-   * @summary Returns the [`Db`](http://mongodb.github.io/node-mongodb-native/3.0/api/Db.html) object corresponding to this collection's database connection from the [npm `mongodb` driver module](https://www.npmjs.com/package/mongodb) which is wrapped by `Mongo.Collection`.
+   * @summary Returns the [`Db`](http://mongodb.github.io/node-mongodb-native/3.6/api/Db.html) object corresponding to this collection's database connection from the [npm `mongodb` driver module](https://www.npmjs.com/package/mongodb) which is wrapped by `Mongo.Collection`.
    * @locus Server
    * @memberof Mongo.Collection
    * @instance